requirements.txt
game-sdk
twitter-plugin-gamesdk
openai
python-dotenv
